Refacor: move review.domain pakage location
