+++ src/main/java/com/cheffi/review/domain/Food.java
@@ -1,4 +1,4 @@
-package com.cheffi.common.config.review.domain;
+package com.cheffi.review.domain;
 
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
@@ -10,29 +10,27 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-public class Review {
+public class Food {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @NotNull
-    private String title;
+    private String name;
 
     @NotNull
-    private String text;
-
-    @NotNull
-    private int ratingCnt;
+    private int price;
 
     @ManyToOne
-    @Column(name = "restuarnt_id")
-    private Restuarant restuarant;
+    @JoinColumn(name = "review_id")
+    private Review review;
+
 
     @Builder
-    public Review(String title, String text, int ratingCnt) {
-        this.title = title;
-        this.text = text;
-        this.ratingCnt = ratingCnt;
+    public Food(String name, int price, Review review) {
+        this.name = name;
+        this.price = price;
+        this.review = review;
     }
 }
